Remove Scala version-specific source directories from core.
The code that adds Scala version-specific unmanaged source directories
seems to be a dead code. The added directories are not in the repository.

The main motivation for removing this code is adding support for Scala
2.11 which is not listed in the pattern match. One way to solve this
problem is to simply add another entry for Scala 2.11 but then the same
problem would reappear when Scala team starts to release Scala 2.12
milestones.