Fix #537 - Deserialization of default values fails if one list has a default value
I changed the way the constructor is selected. If we have a case class, it has two constructors, the normal constructor and the companion apply.

Default values only exists where they are declared, so if apply is selected as constructor, default values don't work.

Now, in case of draw, it chosees the constructor with more defaults (case class constructor over apply constructor).

The problem now is that this code get a bit messy when alternative apply/this() constructors are declared.